+++ src/components/ErrorPopup/ErrorPopup.js
@@ -11,17 +11,6 @@ ErrorPopup.propTypes = {
 
 function ErrorPopup(props) {
   //Компонент вывода ошибки в случае ее возникновения при работе с api
-  function closeByEsc(evt) {
-    if (evt.keyCode === 27) {
-      props.onClose();
-    }
-  }
-  React.useEffect(() => {
-    document.addEventListener("keydown", closeByEsc, false);
-    return () => {
-      document.removeEventListener("keydown", closeByEsc, false);
-    };
-  }, []);
 
   return (
     <section
